+++ rockylinux-server-postinstall.sh
@@ -8,6 +8,8 @@ addsoftwares=(vim tar git tmux ncdu htop rsync git-extras)
 disablefirewalld=0
 # enforcing/permissive/disabled : Statut de SELinux à activer
 selinux=enforcing
+# 
+# postgresql=0
 
 arch=$(uname -m)
 
@@ -99,6 +101,31 @@ if [ ! -f /usr/local/go/bin/go ]; then
 	*) ;;
 	esac
 	echo "Add /usr/local/go/bin to the PATH environment variable"
+fi
+
+if [ ! "vm.swappiness = 1" = "$(sudo sysctl vm.swappiness)" ]; then
+	echo "vm.swappiness = 1" >>/etc/sysctl.d/01-swapp.conf
+	sysctl --load /etc/sysctl.d/01-swapp.conf
+fi
+
+if [ ! "net.ipv6.conf.all.disable_ipv6 = 1" = "$(sudo sysctl net.ipv6.conf.all.disable_ipv6)" ]; then
+	# désactivation de ipv6 pour toutes les interfaces
+	# désactivation de l’auto configuration pour toutes les interfaces
+	echo "net.ipv6.conf.all.disable_ipv6 = 1" >>/etc/sysctl.d/02-ipv6-disable.conf
+	echo "net.ipv6.conf.all.autoconf = 0" >>/etc/sysctl.d/02-ipv6-disable.conf
+	sysctl --load /etc/sysctl.d/02-ipv6-disable.conf
+fi
+if [ ! "net.ipv6.conf.default.disable_ipv6 = 1" = "$(sudo net.ipv6.conf.default.disable_ipv6)" ]; then
+	# désactivation de ipv6 pour les nouvelles interfaces (ex:si ajout de carte réseau)
+	# désactivation de l’auto configuration pour les nouvelles interfaces
+	echo "net.ipv6.conf.default.disable_ipv6 = 1" >>/etc/sysctl.d/02-ipv6-disable.conf
+	echo "net.ipv6.conf.default.autoconf = 0" >>/etc/sysctl.d/02-ipv6-disable.conf
+	sysctl --load /etc/sysctl.d/02-ipv6-disable.conf
+fi
+if [ ! "net.ipv6.conf.lo.disable_ipv6 = 1" = "$(sudo sysctl net.ipv6.conf.lo.disable_ipv6)" ]; then
+	echo "net.ipv6.conf.lo.disable_ipv6 = 1" >>/etc/sysctl.d/02-ipv6-disable.conf
+	echo "net.ipv6.conf.lo.autoconf = 0" >>/etc/sysctl.d/02-ipv6-disable.conf
+	sysctl --load /etc/sysctl.d/02-ipv6-disable.conf
 fi
 
 echo "Preparation completed, it is recommended to restart!"