+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- disable ipv6
+- set vm.swappiness to 1
+
+### Removed
+- balenaEtcher
+- Visual Studio Code
 
 ## [0.1.2] - 2021-10-24
 ### Added